Move the shared error screen into its own component
