+++ src/GlobalErrorHandler.js
@@ -1,26 +1,23 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-import { BaseStyles, GU, PublicUrl, useTheme, useViewport } from '@aragon/ui'
 import * as Sentry from '@sentry/browser'
 import GenericError from './components/Error/GenericError'
 import DAONotFoundError from './components/Error/DAONotFoundError'
 import { network } from './environment'
 import { DAONotFound } from './errors'
 import { getSentryDsn, getPackageVersion } from './local-settings'
-import eagleSvg from './assets/eagle.svg'
-import logo from './assets/logo.png'
+import ErrorScreen from './components/Error/ErrorScreen'
 
 const SENTRY_DSN = getSentryDsn()
 const PACKAGE_VERSION = getPackageVersion()
-const EAGLE_DIMENSIONS = [1307, 877]
 
 class GlobalErrorHandler extends React.Component {
   static propTypes = {
     children: PropTypes.node,
-    compact: PropTypes.bool,
-    theme: PropTypes.object,
   }
+
   state = { error: null, errorStack: null }
+
   componentDidCatch(error, errorInfo) {
     this.setState({
       error,
@@ -52,66 +49,25 @@ class GlobalErrorHandler extends React.Component {
     window.location.reload()
   }
   render() {
-    const { theme, children, compact } = this.props
+    const { children } = this.props
     const { error, errorStack } = this.state
 
-    if (!error) {
-      return children
-    }
-
-    return (
-      <PublicUrl.Provider url="/aragon-ui/">
-        <BaseStyles />
-        <div
-          css={`
-            height: 100vh;
-            min-width: ${45 * GU}px;
-            overflow: auto;
-            background: ${theme.background} fixed 50% 100% /
-              ${EAGLE_DIMENSIONS[0]}px ${EAGLE_DIMENSIONS[1]}px no-repeat
-              url(${eagleSvg});
-          `}
-        >
-          <img
-            src={logo}
-            css={`
-              position: absolute;
-              top: ${2 * GU}px;
-              left: ${2 * GU}px;
-              width: ${4.5 * GU}px;
-              height: ${4.25 * GU}px;
-              opacity: ${Number(!compact)};
-              transition: opacity 50ms;
-            `}
-            alt="Aragon"
+    return error ? (
+      <ErrorScreen>
+        {error instanceof DAONotFound ? (
+          <DAONotFoundError dao={error.dao} />
+        ) : (
+          <GenericError
+            detailsTitle={error.message}
+            detailsContent={errorStack}
+            reportCallback={SENTRY_DSN ? this.handleReportClick : null}
           />
-          <div
-            css={`
-              display: flex;
-              justify-content: center;
-              align-items: center;
-              padding: 50px 20px 20px;
-              min-height: 100%;
-            `}
-          >
-            {error instanceof DAONotFound ? (
-              <DAONotFoundError dao={error.dao} />
-            ) : (
-              <GenericError
-                detailsTitle={error.message}
-                detailsContent={errorStack}
-                reportCallback={SENTRY_DSN ? this.handleReportClick : null}
-              />
-            )}
-          </div>
-        </div>
-      </PublicUrl.Provider>
+        )}
+      </ErrorScreen>
+    ) : (
+      children
     )
   }
 }
 
-export default function GlobalErrorHandlerWithHooks(props) {
-  const theme = useTheme()
-  const { width: vw } = useViewport()
-  return <GlobalErrorHandler theme={theme} {...props} compact={vw < 720} />
-}
+export default GlobalErrorHandler
